# openapi/Documentation.yml
openapi: 3.0.0
info:
  title: Go Server API
  description: API for CRUD operations of users in a PostgreSQL database
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: URL to access the database
paths:
  /users:
    get:
      summary: Gets all users
      responses:
        '200':
          description: List of users obtained successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal error from server
  /users/{id}:
    get:
      summary: Get a user by its ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User obtained successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Not valid id
        '404':
          description: User not found
        '500':
          description: Internal error from server
  /users/create:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
        '405':
          description: Not allowed method
        '409':
          description: Username or Email already exists
        '500':
          description: Internal error from server
  /users/update/{id}:
    patch:
      summary: Update a user by its ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Error in the request (not valid id, username or email)
        '404':
          description: User not found
        '405':
          description: Not allowed method
        '500':
          description: Internal error from server
  /users/delete/{id}:
    delete:
      summary: Delete a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: Not valid id
        '405':
          description: Not allowed method
        '500':
          description: Internal error from server
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time